[fix] rename folders
